Separates beschliessen eines AgendaItems

+++ opengever/meeting/browser/meetings/meeting.py
@@ -61,7 +61,7 @@ AGENDAITEMS_TEMPLATE = '''
 <script id="agendaitemsTemplate" type="text/x-handlebars-template">
   {{#each agendaitems}}
     <tr class="{{css_class}}" data-uid={{id}}>
-      {{#if ../editable}}<td class="sortable-handle"></td>{{/if}}
+      {{#if ../agendalist_editable}}<td class="sortable-handle"></td>{{/if}}
       <td class="number">{{number}}</td>
       <td class="title">
         <span>{{{link}}}</span>
@@ -97,8 +97,13 @@ AGENDAITEMS_TEMPLATE = '''
       {{#if ../editable}}
       <td class="actions">
         <div class="button-group">
-          <a href="{{edit_link}}" title="%(label_edit_action)s" class="button edit-agenda-item"></a>
-          <a href="{{delete_link}}" title="%(label_delete_action)s" class="button delete-agenda-item"></a>
+          {{#if ../agendalist_editable}}
+              <a href="{{edit_link}}" title="%(label_edit_action)s" class="button edit-agenda-item"></a>
+              <a href="{{delete_link}}" title="%(label_delete_action)s" class="button delete-agenda-item"></a>
+          {{/if}}
+          {{#if decide_link}}
+            <a href="{{decide_link}}" title="%(label_decide_action)s" class="button decide-agenda-item"><span></span></a>
+          {{/if}}
         </div>
       </td>
       {{/if}}
@@ -324,10 +329,14 @@ class MeetingView(BrowserView):
         label_edit_save = translate(_('label_edit_save', default='Save'), context=self.request)
         label_edit_action = translate(_('label_edit_action', default='edit title'), context=self.request)
         label_delete_action = translate(_('label_delete_action', default='delete this agenda item'), context=self.request)
+        label_decide_action = translate(
+            _('label_decide_action', default='Decide this agenda item'),
+            context=self.request)
         return AGENDAITEMS_TEMPLATE  % {'label_edit_cancel': label_edit_cancel,
                                         'label_edit_save': label_edit_save,
                                         'label_edit_action': label_edit_action,
-                                        'label_delete_action': label_delete_action}
+                                        'label_delete_action': label_delete_action,
+                                        'label_decide_action': label_decide_action}
 
     def render_handlebars_proposals_template(self):
         label_schedule = translate(_('label_schedule', default='Schedule'), context=self.request)
@@ -337,6 +346,9 @@ class MeetingView(BrowserView):
 
     def json_is_editable(self):
         return json.dumps(self.model.is_editable());
+
+    def json_is_agendalist_editable(self):
+        return json.dumps(self.model.is_agendalist_editable());
 
     @property
     def url_update_agenda_item_order(self):
